1. Add 2 classes in order to divide the burn and migrate responsibilities. 2. Add unit tests(still more to add)
