--- migration/src/test/kotlin/kin/sdk/migration/MigrationManagerTest.kt
@@ -1,37 +1,0 @@
-package kin.sdk.migration
-
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-
-
-class MigrationManagerTest {
-
-    @Before
-    @Throws(Exception::class)
-    fun setUp() {
-
-    }
-
-    @After
-    @Throws(Exception::class)
-    fun tearDown() {
-
-    }
-
-    @Test
-    fun enableLogs() {
-
-    }
-
-    @Test
-    fun getCurrentKinClient() {
-
-    }
-
-    @Test
-    fun start() {
-
-    }
-
-}
